==> Lessons/03_DataExploration_Part2_CEO.Rmd <==
---
title: "5: Data Exploration"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2022"
output: pdf_document
geometry: margin=2.54cm
fig_width: 5
fig_height: 2.5
editor_options: 
  chunk_output_type: console
---
## Remember at End and Start of Sessions
Commit changes, push, when end a session and at beginning if necessary. then pull. If you pull down new changes at the sart of a session remember to push those right after you pull. Pull then push makes sure the changes are on your local machine and on your forked GIT repository online 

## The Outline Icon
The upper right corner of this window has an icon with lines stacked i.e. and outline. Click this icon to see an outline of the rmd. Double hashtags are the headings

## Objectives
1. Import and explore datasets in R
2. Graphically explore datasets in R
3. Apply data exploration skills to a real-world example dataset

## Opening discussion: why do we explore our data?

Why is data exploration our first step in analyzing a dataset? What information do we gain? How does data exploration aid in our decision-making for data analysis steps further down the pipeline?

## Import data and view summaries

```{r, message = FALSE}
# 1. Set up your working directory
getwd()

# 2. Load packages
library(tidyverse)

# 3. Import datasets
USGS.flow.data <- read.csv("../Data/Processed/USGS_Site02085000_Flow_Processed.csv")

#View(USGS.flow.data)
# Alternate option: click on data frame in Environment tab
#check out date to: see if it is what you want/need; determine potential biases; know what and how it's being measured
#numeric data can be discrete or continuous
#categorical are usually character, numbers can be categorical if it is for example a zip code or site ID, those are nominal data meaning the numbers stand for something and are not values to be added up and divided

colnames(USGS.flow.data)
str(USGS.flow.data)
dim(USGS.flow.data)

# Check our date column; remember to check your dates! csv files don't save dates as dates
class(USGS.flow.data$datetime)

USGS.flow.data$datetime <- as.Date(USGS.flow.data$datetime, format = "%Y-%m-%d") 
class(USGS.flow.data$datetime)

```


## Visualization for Data Exploration

Although the `summary()` function is helpful in getting an idea of the spread of values in a numeric dataset, it can be useful to create visual representations of the data to help form hypotheses and direct downstream data analysis. Below is a summary of the useful types of graphs for data exploration.

Note: each of these approaches utilize the package "ggplot2". We will be covering the syntax of ggplot in a later lesson, but for now you should familiarize yourself with the functionality of what each command is doing.

### Bar Chart (function: geom_bar)
gg stands for grammar of graphics; lol love how I'm learning this now after years of using ggplot
aes = aesthetics 

Visualize count data for categorical variables. 
Plots are the dataset and what we want to do with it
bar plots are useful for categorical data for example here it shows what data is approved or provisional or estimated. We provide a column with categorical data in the ggplot code

```{r, fig.height = 3, fig.width = 4}
ggplot(USGS.flow.data, aes(x = discharge.mean.approval)) +
  geom_bar()

#runs a ggplot like above but takes out all the NAs
ggplot(na.omit(USGS.flow.data), aes(x = discharge.mean.approval)) +
  geom_bar()
```

### Histogram (function: geom_histogram)

Visualize distributions of values for continuous numerical variables. What is happening in each line of code? Insert a comment above each line.
Good for looking at numeric data b/c we provide counts by value intervals.

```{r, fig.height = 3, fig.width = 4}
#removes non-finite values i.e. the NAs; double check this is what the warning means

#returns on x-axis the range of values in discharge.mean column
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean))

#binwidth assigns the width of the interval aka bins to the units used in your data column in this case 10 units
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean), binwidth = 10)

#can also just assign the number of bins you want
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean), bins = 20)

#the default bin width is 30, when go from 30 to 20 the bins get fewer and fatter
#generally you use between 5-20 bins 
#too few and don't see trends and distributions
#too many and show the noise and not the signal of the data

#scale_x_continuous allows you to create the scale on your x axis
ggplot(USGS.flow.data, aes(x = discharge.mean)) +
  geom_histogram(binwidth = 10) + 
  scale_x_continuous(limits = c(0, 500))

#shows gage.height.mean instead of the discharge.mean
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = gage.height.mean))


```
### Frequency line graph (function: geom_freqpoly)
Find ggplot cheat sheet under Help (at the top of R window) then cheat sheets

An alternate to a histogram is a frequency polygon graph (distributions of values for continuous numerical variables). Instead of displaying bars,  counts of continuous variables are displayed as lines. This is advantageous if you want to display multiple variables or categories of variables at once.

These are helpful to compare the frequency of things in data. Graph shows us distribution of frequencies not the actual values of the data, so it's okay the mean peak is above the max peak in the graph below

```{r, fig.height = 3, fig.width = 4}
#lty is short for line type
ggplot(USGS.flow.data) +
  geom_freqpoly(aes(x = gage.height.mean), bins = 50) +
  geom_freqpoly(aes(x = gage.height.min), bins = 50, color = "red") +
  geom_freqpoly(aes(x = gage.height.max), bins = 50,  lty = 2) +
  scale_x_continuous(limits = c(0, 10))

#histogram and freqpoly together
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = gage.height.mean), bins = 50) +
geom_freqpoly(aes(x = gage.height.mean), bins = 50) +
  geom_freqpoly(aes(x = gage.height.min), bins = 50, color = "darkgray") +
  geom_freqpoly(aes(x = gage.height.max), bins = 50,  lty = 2) +
  scale_x_continuous(limits = c(0, 10))

#color values by the data in the column after the = sign. In the following example get the A and P in the legend from the gage.height.mean.approval. Can set color to categorical data and show different categories of data on the frequency graph
ggplot(USGS.flow.data) +
  geom_freqpoly(aes(x = gage.height.mean, color = gage.height.mean.approval), bins = 50) +
  scale_x_continuous(limits = c(0, 10)) +
  theme(legend.position = "top")

```
### Box-and-whisker plots (function: geom_boxplot, geom_violin)
Use to see the number of outliers in data and how medians compare across categories

A box-and-whisker plot is yet another alternative to histograms (distributions of values for continuous numerical variables). These plots consist of: 

* A box from the 25th to the 75th percentile of the data, called the interquartile range (IQR).

* A bold line inside the box representing the median value of the data. Whether the median is in the center or off to one side of the IQR will give you an idea about the skewness of your data.

* A line outside of the box representing values falling within 1.5 times the IQR. 

* Points representing outliers, values that fall outside 1.5 times the IQR. 

An alternate option is a violin plot, which displays density distributions, somewhat like a hybrid of the box-and-whiskers and the frequency polygon plot.

Bulk of data is in the box and outliers are in the whiskers

```{r, fig.height = 3, fig.width = 4}
#compress data into box and whiskers. Fat horizontal line is the median. show distribution of mean gage height records.
#Analog to bar plot except looking at distributions instead of counts
ggplot(USGS.flow.data) +
  geom_boxplot(aes(x = gage.height.mean.approval, y = gage.height.mean))

#show distribution in the gage height mean; analogous to histogram where take slice of data and look at distribution by the slice
#low level og gage height we have tight range while larger range at bigger heights
ggplot(USGS.flow.data) +
  geom_boxplot(aes(x = gage.height.mean, y = discharge.mean, group = cut_width(gage.height.mean, 1)))

#violin plot shows distribution within the "box". wider means more data in there.
ggplot(USGS.flow.data) +
  geom_violin(aes(x = gage.height.mean.approval, y = gage.height.mean), 
              draw_quantiles = c(0.25, 0.5, 0.75))
```

### Scatterplot (function: geom_point)
Visualize relationships between continuous numerical variables.
Allows us to explore relations among variables not the distribtutions 

```{r, fig.height = 3, fig.width = 4}
#if the variables below measured the same thing then there'd be a straight diagonal line
#shows relationship between discharge and gage. can see discharge mean doesn't climb as rapidly as gage height and gage height doesn't increase as much as volume increases. stream channel is kinda a bowl and volume increases. makes sense gage.height decreases as stream widens
ggplot(USGS.flow.data) +
  geom_point(aes(x = discharge.mean, y = gage.height.mean))

#date.time is hybrid between categorical and continuous
#showing mean discharge at different times
#if x is set as a purely categorical variable it's not as interesting b/c then the data will all just line up
ggplot(USGS.flow.data) +
  geom_point(aes(x = datetime, y = discharge.mean))
```

Question: under what circumstances would it be beneficial to use each of these graph types (bar plot, histogram, frequency polygon, box-and whisker, violin, scatterplot)?

> Answer: 

## Ending discussion

What did you learn about the USGS discharge dataset today? What separate insights did the different graph types offer?
> Answer: 


How can multiple options for data exploration inform our understanding of our data?

> Answer: 


Do you see any patterns in the USGS data for the Eno River? What might be responsible for those patterns and/or relationships?

> Answer:
