==> Lessons/04_Lab_DataWrangling.Rmd <==
---
title: "Lab 4: Data Wrangling"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1- Answer questions on M3/A3
2- Answer questions on M4
3- Practice wrangling datasets with dplyr functions: use functions and think about what it is actually doing

## Set up your session

Today we will work with a dataset from the [North Temperate Lakes Long-Term Ecological Research Station](https://lter.limnology.wisc.edu/about/overview). The NTL-LTER is located in the boreal zone in northern Wisconsin, USA. We will use the [chemical and physical limnology dataset](https://lter.limnology.wisc.edu/content/cascade-project-north-temperate-lakes-lter-core-data-physical-and-chemical-limnology-1984), running from 1984-2016. 

Opening discussion: why might we be interested in long-term observations of temperature, oxygen, and light in lakes?

> Add notes here: 

```{r, message = FALSE}
getwd()
#install.packages(tidyverse)
library(tidyverse)
#install.packages(lubridate)
library(lubridate)
NTL.phys.data <- read.csv("./Data/Raw/NTL-LTER_Lake_ChemistryPhysics_Raw.csv", stringsAsFactors = TRUE)

class(NTL.phys.data$sampledate)
# Format sampledate as date
NTL.phys.data$sampledate <- as.Date(NTL.phys.data$sampledate, format = "%m/%d/%y")
```


## Filter

Filtering allows us to choose certain rows (observations) in our dataset.

```{r}
#when filtering it is important to know what class you're working with so that you know how to run the filter i.e. does it need quotes around it
class(NTL.phys.data$lakeid)
class(NTL.phys.data$depth)

# dplyr filtering. A good check to see if filtering did something is to check that the filtered dataset has fewer observation than the unfiltered
NTL.phys.data.surface <- filter(NTL.phys.data, depth == 0)

# Choose multiple conditions to filter
summary(NTL.phys.data$lakename)
NTL.phys.data.PeterPaul <- filter(NTL.phys.data, lakename %in% c("Paul Lake", "Peter Lake"))

# Choose a range of conditions of a numeric or integer variable
summary(NTL.phys.data$daynum)
NTL.phys.data.JunethruOctober <- filter(NTL.phys.data, daynum %in% c(152:304)) #c creates a range/sequence of numners you need
#152:304 represents the day of the year that are in the desired months hence the daynum (which is a column in the dataset) - double check the daynumb in your your data b/c some things coud've mixed up
#use daynum b/c want June-Oct for all years of our dataset
#daynum is not a function in R. It is a column in this dataset

# Exercise 1: 
# filter NTL.phys.data for the year 1999
# what code do you need to use, based on the class of the variable?
class(NTL.phys.data$year4) #integer
NTL.phys.data.1999 <- filter(NTL.phys.data, year4 %in% 1999)

#also could do it this way
NTL.phys.data.1999 <- filter(NTL.phys.data, year4 == 1999)

summary(NTL.phys.data.1999)
# Exercise 2: 
# filter NTL.phys.data for Tuesday Lake from 1990 through 1999.
NTL.phys.data.Tuesday <- filter(NTL.phys.data, lakename == "Tuesday Lake")
NTL.phys.data.Tuesday.1990to1999 <- 
  #whoops finish this out!

NTL.phys.data.1990.1999 <- filter(NTL.phys.data, year4 %in% c(1990:1999))

summary(NTL.phys.data.1990.1999)

#remember when filtering the number of column from the original to new datasets should remain the same. Only the number of rows changes

```
Question: Why don't we filter using row numbers?

> Answer: 

## Pipes

Pipe is another method to wrangle datasets that looks cleaner and is easier to read.  We designate a pipe with `%>%`. A good way to think about the function of a pipe is with the word "then." 

Let's say we want to take our raw dataset (NTL.phys.data), *then* filter the data for Peter and Paul lakes, *then* select temperature and observation information, and *then* add a column for temperature in Fahrenheit: 

```{r}
#In pipes order matters. code is done in the order it is written. 
NTL.phys.data.processed <- 
  NTL.phys.data %>%
  filter(lakename == "Paul Lake" | lakename == "Peter Lake") %>%
  select(lakename, sampledate:temperature_C) %>%
  mutate(temperature_F = (temperature_C*9/5) + 32)

#change to Celsius! mutate adds a new column.

#Exercise 3: Using a pipe filter NTL.phys.data for Tuesday Lake from 1990 through 1999 only for July 

NTL.phys.data.Tuesday.July <- 
  NTL.phys.data %>%
  filter(lakename == "Tuesday Lake") %>%
  filter(year4 %in% c(1990:1999)) %>% 
  filter(daynum %in% c(183:213))
#using daynum can cause problems with leap years so you can mutate to get month

NTL.phys.data.Tuesday.July <- 
  NTL.phys.data %>%
  filter(lakename == "Tuesday Lake") %>%
  filter(year4 %in% c(1990:1999)) %>% 
  filter(month(sampledate) == 7)
  

#could mutate but since you don't need the month coulmn then can bypass it
mutate(month = format(sampledate, "%m")) %>% 
  filter(month == 7)

#Exercise 4: Using the data from part 3, a pipe and the summarise() function find the mean surface temperature (hints: you will need to add another filter for depth==0). Make sure you eliminate NAs before computing the means

NTL.phys.data.mean %>% 
  NTL.phys.data %>%
  filter(depth == 0) %>% 
  filter(!is.na(temperature_C)) %>% 
  summarise(mean = mean(temperature_C))
  
#exclamation point means the value is different from like the double equals is equal to
```

## Gather and Spread

For gather we will use `pivot_longer` and for spread we will use `pivot_wider`.

```{r}
#Exercise 5: gather irradiance data (measured in the water column and measured on the deck of the sampling boat) into one column using pivot_longer. Could have 3,4,5etc coulmns into two columns

#The irradianceWater and irradianceDeck are the columns you are combining

NTL.phys.data.gathered <- 
  pivot_longer(NTL.phys.data, irradianceWater:irradianceDeck, names_to = "irradiance", values_to = "radiation")

#Exercise 6: spread temperature_C into more than one column based on the depth

NTL.phys.data.spread <- pivot_wider(NTL.phys.data, names_from = "depth", values_from = "temperature_C")

```
